+++ models/umowy_zlecenie.js
@@ -3,17 +3,17 @@
 module.exports = function(sequelize, DataTypes) {
   return sequelize.define('umowy_zlecenie', {
     IdUmowy: {
-      type: DataTypes.INTEGER(11),
+      type: DataTypes.INTEGER(10),
       allowNull: false,
       primaryKey: true,
       autoIncrement: true
     },
     StatusStudenta: {
-      type: DataTypes.INTEGER(1),
+      type: DataTypes.INTEGER(3),
       allowNull: false
     },
     ZUS: {
-      type: DataTypes.INTEGER(1),
+      type: DataTypes.INTEGER(3),
       allowNull: false
     }
   }, {
